\name{smogs-package}
\alias{smogs-package}
\docType{package}
\title{
Sparse Multitask Overlap Group Selection
}
\description{
This package is designed to do variable selection of groups of features, which are allowed to overlap, in several tasks. The model also supports sparsity within groups so that individual features can be set to zero.}
\details{
\tabular{ll}{
Package: \tab smogs\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-28\cr
License: \tab GPL-3\cr
}
Two main functions; \code{smogs} and \code{sogl} do group selection in multiple, and a single task, respectively.}
\author{
Alexandra Jauhiainen and Karl Landström

Maintainer: Alexandra Jauhiainen <alexandra.jauhiainen@ki.se>, Karl Landström <karl.landstrom@ki.se>
}
\references{
Shotgun C++ package: http://www.select.cs.cmu.edu/projects/shotgun/
}
\keyword{package}
\keyword{variable selection}
\keyword{sparsity}
\keyword{hierarchical lasso}
\keyword{group lasso}
\keyword{multitask}
